refactor(stats): extract reusable stats screens
